Add config option enable_preview_support
When set to false this prevents the preview module being inserted into
all controllers. The preview code returns Rooftop.include_drafts to false
before every action - which makes Rooftop.include_drafts of very limited
use because it does not persist.

Default it to true to maintain same behaviour as previously.